+++ local_planner/src/nodes/common.h
@@ -37,7 +37,7 @@ float distance2DPolar(const PolarPoint& p1, const PolarPoint& p2);
 **/
 Eigen::Vector3f polarToCartesian(const PolarPoint& p_pol,
                                  const geometry_msgs::Point& pos);
-double indexAngleDifference(float a, float b);
+float indexAngleDifference(float a, float b);
 /**
 * @brief     compute point in the histogram to a polar point
 * @param[in] e evelation index in the histogram
