+++ local_planner/src/nodes/local_planner.cpp
@@ -18,7 +18,7 @@ void LocalPlanner::setPose(const geometry_msgs::PoseStamped msg) {
   pose_.header = msg.header;
   pose_.pose.position = msg.pose.position;
   pose_.pose.orientation = msg.pose.orientation;
-  curr_yaw_ = tf::getYaw(msg.pose.orientation);
+  curr_yaw_ = static_cast<float>(tf::getYaw(msg.pose.orientation));
   star_planner_->setPose(pose_, curr_yaw_);
 
   if (!currently_armed_ && !disable_rise_to_goal_altitude_) {
